# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dead Pigeons API",
    "description": "API for the Dead Pigeons bingo project.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "Admin_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "Admin_Create",
        "requestBody": {
          "x-name": "admin",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Admin"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Admin/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "Admin_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "Admin_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedAdmin",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Admin"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "operationId": "Admin_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Board": {
      "get": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Create",
        "requestBody": {
          "x-name": "board",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Board/{id}": {
      "get": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedBoard",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Board"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game": {
      "get": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_Create",
        "requestBody": {
          "x-name": "game",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{id}": {
      "get": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedGame",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Player": {
      "get": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Create",
        "requestBody": {
          "x-name": "player",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Player/{id}": {
      "get": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedPlayer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Player"
        ],
        "operationId": "Player_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Admin": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "adminId": {
            "type": "string",
            "format": "guid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          }
        }
      },
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "boardId": {
            "type": "string",
            "format": "guid"
          },
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "chosenNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "isRepeating": {
            "type": "boolean"
          },
          "repeatUntilGameId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          },
          "repeatUntilGame": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Game"
              }
            ]
          },
          "winningboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Winningboard"
            }
          }
        }
      },
      "Game": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "winningNumbers": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "drawDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "boardGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "boardRepeatUntilGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "winningboards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Winningboard"
            }
          }
        }
      },
      "Winningboard": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "winningboardId": {
            "type": "string",
            "format": "guid"
          },
          "gameId": {
            "type": "string",
            "format": "guid"
          },
          "boardId": {
            "type": "string",
            "format": "guid"
          },
          "winningNumbersMatched": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "board": {
            "$ref": "#/components/schemas/Board"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          }
        }
      },
      "Player": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "guid"
          },
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "mobilepayReqId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          }
        }
      }
    }
  }
}